Created evioUtil<T>::getContentType() to solve solaris bug

+++ swap_util.c
@@ -18,6 +18,9 @@
  *
  * Revision History:
  *   $Log$
+ *   Revision 1.8  2007/01/09 17:56:30  wolin
+ *   Created evioUtil<T>::getContentType() to solve solaris bug
+ *
  *   Revision 1.7  2007/01/09 16:09:26  wolin
  *   Minor mods for solaris
  *
@@ -489,7 +492,7 @@ void swapped_memcpy(char *buffer,char *source,int size)
 	sg_tag  = (header2 >> 24) & (0xff);
 	sg_type = (header2 >> 16) & (0xff);
 	if((sg_type>=0x10)||(sg_type==0xc)||(sg_type==0xd)||(sg_type==0xe)) {
-	  evStack_pushon((sg_size)*2,i,sg_type,sg_tag,NULL,head);
+	  evStack_pushon((sg_size)*2,i,sg_type,sg_tag,(int)NULL,head);
 	  lk.head_pos = i + 2;
 	  head->length += 1;
 	  i = i+ 2;
@@ -512,7 +515,7 @@ void swapped_memcpy(char *buffer,char *source,int size)
 	tsg_tag  = (header2 >> 20) & (0xfff);
 	tsg_type = (header2 >> 16) & (0xf);
 	if((tsg_type==0xc)||(tsg_type==0xd)||(tsg_type==0xe)) {
-	  evStack_pushon((tsg_size)*2,i,tsg_type,tsg_tag,NULL,head);
+	  evStack_pushon((tsg_size)*2,i,tsg_type,tsg_tag,(int)NULL,head);
 	  lk.head_pos = i + 2;
 	  head->length += 1;
 	  i = i+ 2;
